use common functions for managing client apps, and update for npm5+

+++ src/client_apps/_vars.sh
@@ -36,8 +36,15 @@ install_ngcli_global() {
 
 install_libs() {
     cd "${1}"
-    rm -rf node_modules
-    npm install
+
+    if [ -e 'package-lock.json' ]
+    then
+        npm ci
+    else
+        rm -rf node_modules
+        npm install
+    fi
+
     cd ..
 }
